requirements.txt
google-generativeai
requests
rich
docker
gitpython
python-dotenv
